pyproject.toml
[project]
name = "lifesight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.136.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "markdown>=3.10.2",
    "pydantic-settings>=2.14.0",
    "pypdf>=6.10.2",
    "python-multipart>=0.0.26",
    "uvicorn>=0.46.0",
]
